add outlogin and update login

+++ src/main/java/com/xcy/community/dto/QuestionDto.java
@@ -8,7 +8,7 @@ public class QuestionDto {
     private String description;
     private Long gmtCreat;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
@@ -55,11 +55,11 @@ public class QuestionDto {
         this.gmtModified = gmtModified;
     }
 
-    public Integer getCreator() {
+    public Long getCreator() {
         return creator;
     }
 
-    public void setCreator(Integer creator) {
+    public void setCreator(Long creator) {
         this.creator = creator;
     }
